networksecurity/utils/main_utils/utils.py
from networksecurity.exception.exception import NetworkSecurityException
from networksecurity.logging.logger import logging
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import r2_score
import yaml
# import dill 
import numpy as np
import pickle
import sys
import os


def read_yaml_file(filepath: str) -> dict:
    try:
        with open(filepath, 'r') as yaml_file:
            return yaml.safe_load(yaml_file)
    except Exception as e:
        raise NetworkSecurityException(e, sys)

def write_yaml_file(filepath : str , content : object , replace:bool = False )->None:
    try:
        if replace:
            if os.path.exists(filepath):
                os.remove(filepath)
        
        os.makedirs(os.path.dirname(filepath),exist_ok=True)
        with open(filepath,'w') as file :
            yaml.dump(content,file)
    except Exception as e:
        raise NetworkSecurityException(e,sys)